requirements.txt
fastapifastapi
uvicorn
pydantic-settings
